requirements.txt
langchain==0.1.20
langchain-openai
langchain-community
python-dotenv
pydantic
wikipedia
